+++ datafusion/core/src/datasource/listing_table_factory.rs
@@ -838,7 +838,10 @@ mod tests {
         use datafusion_execution::config::SessionConfig;
         use datafusion_physical_expr::PhysicalExpr;
         use datafusion_physical_plan::ExecutionPlan;
-        use datafusion_session::{CatalogProviderList, EmptyCatalogProviderList};
+        use datafusion_session::{
+            CatalogProviderList, EmptyCatalogProviderList, QueryPlanner,
+            UnsupportedQueryPlanner,
+        };
         use std::any::Any;
         use std::collections::HashMap;
 
@@ -856,6 +859,9 @@ mod tests {
             }
             fn catalog_list(&self) -> Arc<dyn CatalogProviderList> {
                 Arc::new(EmptyCatalogProviderList)
+            }
+            fn query_planner(&self) -> Arc<dyn QueryPlanner + Send + Sync> {
+                Arc::new(UnsupportedQueryPlanner)
             }
             async fn create_physical_plan(
                 &self,
