+++ docs/source/library-user-guide/upgrading/55.0.0.md
@@ -823,18 +823,29 @@ fn catalog_list(&self) -> Arc<dyn CatalogProviderList> {
 }
 ```
 
-`Session` also provides `query_planner`, `optimize`, `physical_optimizers`, and
-`statistics_registry`. The default implementations report that the query
-planner is unavailable, leave logical plans unchanged, return no physical
+`Session` now also requires a `query_planner` method. Implementations that do
+not expose a query planner can return the new `UnsupportedQueryPlanner`:
+
+```rust
+use std::sync::Arc;
+use datafusion_session::{QueryPlanner, UnsupportedQueryPlanner};
+
+fn query_planner(&self) -> Arc<dyn QueryPlanner + Send + Sync> {
+    Arc::new(UnsupportedQueryPlanner)
+}
+```
+
+The `optimize`, `physical_optimizers`, and `statistics_registry` methods have
+default implementations that leave logical plans unchanged, return no physical
 optimizer rules, and return no statistics registry, respectively. A custom
 session that uses `DefaultQueryPlanner` or `DefaultPhysicalPlanner` should
 override these methods to expose its planning and optimization behavior.
 
 `ForeignSession::create_physical_plan` continues to run the complete planning
-pipeline in the library that owns the session. Directly passing a
-`ForeignSession` to a local `DefaultPhysicalPlanner` uses the default local
-optimization capabilities described above. FFI wrappers for the individual
-planner and optimizer interfaces are not included in this release.
+pipeline in the library that owns the session. `ForeignSession::query_planner`
+returns `UnsupportedQueryPlanner` until the query planner FFI interface is
+available. FFI wrappers for the individual planner and optimizer interfaces are
+not included in this release.
 
 See [PR #23703](https://github.com/apache/datafusion/pull/23703) for details on
 the catalog changes.
